@@ -1,5 +1,5 @@
 module moderngo-testdata
 
-go 1.25.0
+go 1.26.0
 
 require github.com/quasilyte/go-ruleguard/dsl v0.3.23 // indirect
